requirements.txt
annotated-types==0.6.0
anyio==4.2.0
argcomplete==3.1.6
beautifulsoup4==4.12.3
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cssselect==1.2.0
distro==1.9.0
dnspython==2.4.2
email-validator==2.1.0.post1
fastapi==0.109.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.13.1
fsspec==2023.12.2
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.20.3
idna==3.6
itsdangerous==2.1.2
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.3.2
lxml==5.1.0
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
newsapi-python==0.2.7
newsdataapi==0.1.10
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.3
openai==1.10.0
orjson==3.9.10
packaging==23.2
pillow==10.2.0
pipx==1.2.1
pydantic==2.6.0
pydantic-extra-types==2.1.0
pydantic-settings==2.1.0
pydantic_core==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.6
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-file==2.0.0
safetensors==0.4.2
scikit-learn==1.4.0
scipy==1.12.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.35.1
sympy==1.12
threadpoolctl==3.2.0
tinysegmenter==0.3
tldextract==5.1.1
tokenizers==0.15.1
torch==2.2.0
tqdm==4.66.1
transformers==4.37.2
typing_extensions==4.9.0
ujson==5.8.0
urllib3==2.1.0
userpath==1.9.1
uvicorn==0.24.0.post1
watchfiles==0.21.0
websockets==12.0
